+++ astropy/coordinates/tests/test_frames_with_velocity.py
@@ -281,3 +281,23 @@ def test_shorthand_attributes():
                  representation_type=r.CartesianRepresentation,
                  differential_type=r.CartesianDifferential)
     icrs4.radial_velocity
+
+
+def test_negative_distance():
+    """ Regression test: #7408
+    Make sure that negative parallaxes turned into distances are handled right
+    """
+
+    RA = 150 * u.deg
+    DEC = -11*u.deg
+    c = ICRS(ra=RA, dec=DEC,
+             distance=(-10*u.mas).to(u.pc, u.parallax()),
+             pm_ra_cosdec=10*u.mas/u.yr,
+             pm_dec=10*u.mas/u.yr)
+    assert quantity_allclose(c.ra, RA)
+    assert quantity_allclose(c.dec, DEC)
+
+    c = ICRS(ra=RA, dec=DEC,
+             distance=(-10*u.mas).to(u.pc, u.parallax()))
+    assert quantity_allclose(c.ra, RA)
+    assert quantity_allclose(c.dec, DEC)